Refactor of the kernel method is WORKING
